# openapi-shopiea.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Score API",
    "description": "shopiea is scoring api to manage user score, class, course, lab, export to csv, etc.",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:9898/v1"
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "User Authentication"
    },
    {
      "name": "users",
      "description": "User operations"
    },
    {
      "name": "class",
      "description": "class operations"
    },
    {
      "name": "courses",
      "description": "Course operations"
    },
    {
      "name": "labs",
      "description": "Lab operations"
    },
    {
      "name": "export",
      "description": "Export to CSV"
    },
    {
      "name": "score",
      "description": "Score operations"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "summary": "Login",
        "description": "Login with username and password",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "301": {
            "description": "Invalid username/password supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralError"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all users by class",
        "description": "Get users",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "class",
            "in": "query",
            "description": "Class name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "class1"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "404": {
            "description": "course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralError"
                }
              }
            }
          }
        }
      }
    },
    "/classes": {
      "get": {
        "tags": [
          "class"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get classes",
        "description": "Get classes",
        "operationId": "getClasses",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "class"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create class",
        "description": "Create class",
        "operationId": "createClass",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClassResponse"
                }
              }
            }
          }
        }
      }
    },
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get courses",
        "description": "Get courses",
        "operationId": "getCourses",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoursesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Submit course",
        "description": "Submit course",
        "operationId": "submitCourse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCourseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/labs": {
      "get": {
        "tags": [
          "labs"
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "description": "Course name",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": "1"
          }
        ],
        "summary": "Get labs",
        "description": "Get labs",
        "operationId": "getLabs",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "labs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Submit lab",
        "description": "Submit lab",
        "operationId": "submitLab",
        "parameters": [
          {
            "name": "course",
            "in": "path",
            "description": "Course name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "course1"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLabResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scores/{course}": {
      "get": {
        "tags": [
          "score"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get scores",
        "description": "Get scores",
        "operationId": "getScores",
        "parameters": [
          {
            "name": "course",
            "in": "path",
            "description": "Course name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "course1"
          },
          {
            "name": "lab",
            "in": "query",
            "description": "Lab name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "lab1"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoresResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "score"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Post scores",
        "description": "Post scores",
        "operationId": "postScores",
        "parameters": [
          {
            "name": "course",
            "in": "path",
            "description": "Course name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "course1"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushScoreResponse"
                }
              }
            }
          }
        }
      }
    },
    "/export": {
      "get": {
        "tags": [
          "export"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Export score",
        "description": "Export score",
        "operationId": "exportScore",
        "parameters": [
          {
            "name": "course",
            "in": "query",
            "description": "Course name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "course1"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GeneralError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expire": {
            "type": "string"
          }
        }
      },
      "UsersResponse": {
        "type": "object",
        "properties": {
          "class": {
            "type": "string"
          },
          "students": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ClassesResponse": {
        "type": "object",
        "properties": {
          "class": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateClassRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateClassResponse": {
        "type": "object",
        "properties": {
          "class": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "LabsResponse": {
        "type": "object",
        "properties": {
          "labs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateLabRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateLabResponse": {
        "type": "object",
        "properties": {
          "lab": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CoursesResponse": {
        "type": "object",
        "properties": {
          "courses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateCourseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateCourseResponse": {
        "type": "object",
        "properties": {
          "course": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ScoresResponse": {
        "type": "object",
        "properties": {
          "scores": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "lab": {
                  "type": "string"
                },
                "score": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "PushScoreRequest": {
        "type": "object",
        "properties": {
          "lab": {
            "type": "string"
          },
          "score": {
            "type": "integer"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "PushScoreResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ExportResponse": {
        "type": "object",
        "properties": {
          "course": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "report": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "avarage": {
                  "type": "integer"
                },
                "scores": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "lab": {
                        "type": "string"
                      },
                      "score": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}